+++ engine/dlib/src/dlib/trig_lookup.h
@@ -47,7 +47,7 @@ namespace dmTrigLookup
      */
     inline float Sin(float radians)
     {
-        return Cos(radians - M_PI_2);
+        return Cos(radians - (float) M_PI_2);
     }
 
 }
